Prometheus generate-manifest.sh adds alerts.d ops
We want to be able to add new custom alerts to prometheus.

To do so, we can set the variable `custom_rules`[1] using opsfiles.

We will read these files using a <dir>.d/*.yml approach, similar
to operations.d, but we keep it separated for clarity.

We also add the posibility to pass a list of VARS_FILES, as we
will likely require additional variables from, for example,
terraform outputs.

[1] https://bosh.io/jobs/prometheus2?source=github.com/cloudfoundry-community/prometheus-boshrelease&version=23.3.0#p%3dprometheus.custom_rules

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>