Add smtp receiver to alertmanager
We use the same credentials than for UAA to send notifications
or reset passwords. These are the creds of a AWS user with access
to SES.

For that we need to consume the terraform outputs for the cf
terraform. We add the variables to the terraform manifest generation
in the pipeline and tests.

Currently we configure to send email to our alerting groups:

- govpaas-alerting-dev@digital.cabinet-office.gov.uk
- govpaas-alerting-staging@digital.cabinet-office.gov.uk
- govpaas-alerting-prod@digital.cabinet-office.gov.uk

The alert would be sent by mail if it includes the word "email"
in the label notify.

We set continue: true so that the alert can be sent to more
receivers.

Signed-off-by: Lee Porte <lee.porte@digital.cabinet-office.gov.uk>