Upload paas-accounts documents from the pipeline
paas-accounts has a database containing documents that users should
agree to when they start using PaaS.

Because the documents are uploaded to the API manually in each
environment the environments don't all match.

Currently the terms of service is present in Ireland, but not London.

This adds the essential documents (that you would expect to exist in
every environment) to paas-accounts through the pipeline.

The API call is a PUT, so it should be safe to call it every time the
pipeline runs (although it will overwrite any edits made through the API
between deployments).