Unbind splitted paas-billing before delete
On #1629 we splitted the paas-billing in two different deployed
apps, so that we can run the API in HA and the collectors as
single instance.

But the autodelete-deployment needs to unbind all the apps to
delete the service, and now we have two apps bound to the same
service.

We extend the remove-db.yml task to accept a list of BOUND_APPS
to unbind, instead of only one app.