[paas-metrics] Create a Prometheus scrapable metrics endpoint
We want to allow Prometheus to scrape the same metrics we are currently sending
to Datadog. To allow this we created a multiwriter which can send events to
multiple reporters.

For testing purposes you can set DEBUG=1 which will add and stdout reporter to
the multiwriter.

You can disable the Datadog reporter by setting DISABLE_DATADOG=1.