Add job to rotate prometheus credentials
We want to be able to rotate the prometheus secrets generated
by bosh during the prometheus deployment.

As we store the secrets in a vars-store.yml file,
we reuse the script used for the cf-manifest:

   manifests/cf-manifest/scripts/rotate-vars-store-secrets.rb

This script requires a version of the manifest without the variables
expanded, so we update the render-manifest.sh script to be able
to render the manifest without variables, and we pass the
right arguments as needed.

The rendered manifest without variables is stored as a resource
in the S3 bucket.

Note that currently the prometheus-vars-store only contains a set
of plain passwords, we do not use the `*_old` vars, and we do not
keep any variable. That is effectivelly like deleting all the
vars-store.yml file. But we decided to use the
rotate-vars-store-secrets.rb script as it is more flexible.