Refactor release version manifest test
We want to make it easier to pin specific versions of releases for
given versions of upstream releases.

This covers the cases of pin to an older version or to a custom version.

We want the test to fail if the upstream version changes, so we
double check if we still have to pin that release.

In this implementation, the developer only needs to add a new
entry to the variable pinned_releases, with the name, local version
and upstream version of the release.