Alert on average total diego-cell memory used
In `528417be` we record a metric and alert on it per-diego-cell, but we
actually want to alert on the average *aggregate* memory that the cells
have free

We can do this by averaging a new derived metric (per-VM, recorded for
all VMs regardless of purpose) and comparing the percentage for our
diego cells with 50% and 55% thresholds.

NB This is only valid as long as our diego cells are (memory-)
homogenous!