Set data retention to 450 days and increase the persistent storage for Prometheus to 100Gb
We decided to have a data retention period of 15 months, so we can look back at
least a year (and some).
Based on our current usage we store 120Mb/day, which would be 54G for 15
months. We want to leave some room for growing, so we'll attach a 100Gb disk as
data storage.