Change Prometheus max-block-duration to 7d
Some Prometheus users reported high CPU and memory usage if high retention
periods are used (e.g. >1 year), and the max block duration is not set (which
defaults to 10% of the retention period).

According to [1] we should have a max block duration around 1-2% of the
retention period, therefore we set it to 7 days.

[1] https://github.com/prometheus/prometheus/issues/4110