Ignore "overlay" device in disk-space alert
The "overlay" device is for the containers of the applications, and
the space is restricted to the allowed disk for the application.

We do not want to alert if an application runs out of space, that
is up for the tenant to monitor.

Because that, we also ignore the overlay device.

Example output of a df command:

overlay             1024   998        27  98% /var/vcap/data/grootfs/store/unprivileged/images/ab308816-7857-4c76-5020-9114/rootfs