Remove Datadog concourse monitors based on Riemman
We had to stop using the Riemman exporter for concourse[1] as
concourse only supports one metric emitter[2]. This emitter was
being used to get the metrics into Datadog, but now all
the corresponding monitors and dashboard would start failing in
Datadog.

We have migrated them to prometheus[3] so we can delete them now
from Datadog, so they stop alerting.

[1] https://github.com/alphagov/paas-bootstrap/pull/227/commits/b9323815c2a0b32c61e8752ba9969a0380e9987f
[2] https://github.com/concourse/atc/blob/f8bd990b2a64b380651ebb6c5999bbcad10de293/metric/emit.go#L73
[3] https://github.com/alphagov/paas-cf/pull/1681