Fix firing alerts N/A on prometheus dashboard
We originally set up the "Firing alerts" panel to replace "no data" with
"N/A". This is a bit confusing, because really when there's no data it
means "0 alerts". To make matters worse, if the alerts fire the panel
goes red, but the transition from firing to not firing doesn't reset the
color.

Much better to set the panel to 0 if there's no data (which can be done
with `or vector(0)`). I also noticed that the thresholds were wrong (it
was showing as orange when there were 0 alerts).