Remove DataDog part 1: Remove all monitors
In order to remove DataDog without causing loads of noise this runs a
DataDog terraform destroy to remove all the monitors ahead of upcoming
changes to remove the agents etc that generate the metrics.

This also removes the config step to genetate the list of job_instances
that's only used in the datadog terraform apply task.