Parse application logs sent by a syslog drain
We want to send some of our application logs to Logit. The current Logstash
filters can already parse the RTR and APP logs, but that was meant to work if
all app logs are parsed on the platform.

Our only options seems to be to check if the syslog_program is a uuid as that
means the source is a CF application.

For the RTR/WEB log parsing to work we need to parse the syslog_procid (e.g.
"[RTR/0]") and set "RTR" or "WEB/*" to `[parsed_json_field][source_type]`. We
could do a less hacky solution, but then we would need to figure out how to
inject some logstash filters in the middle of the upstream ones.