Allow rotate CC database encryption key
We will implement a logic similar to the one implemented for
the jwt token[1]. The CC encryption key name would be a password itself
that would be rotated, called `cc_db_encryption_key_id`

We use the `cc_db_encryption_key_id_old` and `cc_db_encryption_key_old`
to allow transit between key rotations.

Here we add a dedicated job to rotate the DB keys, as it is a more
sensitive task that will not be run as often. This task rotates
the database key `cc_db_encryption_key` and the id
`cc_db_encryption_key_id` at the same time.

We need to also add cc_db_encryption_key_id as a secret to preserve
during the key rotation.

We move the related configuration of the vm_extension for the errand
definition to the operations file including all the rotation
config.

We drop the the default `encryption_key_0` that comes from
cf-deployment, but we will add logic in a later commit to retrieve the
current value to allow rotation for the first time.

[1] https://github.com/alphagov/paas-cf/pull/1184