Allow rotate UAA database encryption key
We will implement a logic similar to the one implemented for
the jwt token and the CC db encryption key.

The UAA encryption key name would be a password itself
that would be rotated, called `uaa_default_encryption_passphrase_id`

We use the `uaa_default_encryption_passphrase_id_old` and
`uaa_default_encryption_passphrase_old` to allow transit
between key rotations.

We add an errand to run the uaa_key_rotator job.

Here we add the corresponding task to the job to rotate the DB keys.
This task must be run sequencially (under a `- do:` after the
existing CC db key rotation task, because the bosh errad adquires
a global lock on the manifest when running, meaning that only
one errand or deployment can run concurrently.

We need to also add uaa_default_encryption_passphrase_id as a secret
to preserve during the key rotation.

Here we stop using the default `default_key` that comes from
cf-deployment, but we will add logic in a later commit to
keep the current value in the current deployments to allow the
initial key rotation.

[1] https://github.com/alphagov/paas-cf/pull/1184