Keep the old value of default DB encrypyion keys
We are already using the default encryption keys provided by
the cf-deployment manifest for UAA and CC.

Apparently there is some safety check in the CC and UAA jobs[1][2]
that prevent you from changing the values of the encryption keys,
even when the value is not used anymore.

Because that, we must record the old value of:

 - `api.properties.cc.database_encryption.keys.encryption_key_0: "((cc_db_encryption_key))"`
 - `uaa.properties.encryption.encryption_keys.label=default_key.passphrase: ((uaa_default_encryption_passphrase))`

We do so during the manifest interpolation with some bash-fu that
we will delete in the future. We add two variables to the manifest
to keep track of these secrets.

Once we deployed this once, and the passwords have been rotated,
we can revert this commit

[1] https://github.com/cloudfoundry/capi-release/blob/develop/jobs/cloud_controller_ng/templates/validate_encryption_keys.sh.erb