Get bosh admin-password from bosh-vars-store.yml
We have migrated to bosh-deployment based manifests and the secrets
are being generated in bosh-interpolate and stored in the
file bosh-vars-store.yml[1]

We need to replace every point where we pull the bosh admin password
to get it from the bosh-vars-store.yml file.

We also renamed the variable for the prometheus bosh-exporter client.

[1] https://github.com/alphagov/paas-bootstrap/pull/230