Upload buildpacks from post-deploy tasks
We want to be able to upload the buildpacks directly from an
upstream URL, rather than using bosh-releases. That way we do not
depend on upstream releases.

For that we add a new task that would upload every single buildpack
enable. The task calls a helper script `upload-buildpack.sh`
for each buildpack and specifies:

 - name of the buildpack
 - stack (cflinuxfs2, cflinuxfs3)
 - filename (should include version)
 - URL where download the buildpack
 - checksum

We keep all this info in the pipeline task itself, as it is the
easiest and simplest solution.

The helper script would check if the buildpack is already installed
and if not or if the version changed, it will download, check the
checksum and create or update the corresponding buildpack.

For the initial creation we had to implement some additional workaround
for the special case of java. For most of the buildpacks CF would set
the stack automatically, except for java. In that case, in order to
set the stack, we must upload it with a different name to avoid
conflicts with other stacks, set the stack and then rename it.