Disable email alerts on development environments
We want to stop the dev environments from sending email notifications
unless enabled.

We disable the notifications for the smoke tests and for the alert
manager, by adding a new parameter ENABLE_ALERT_EMAILS, that
is true by default and set to false in dev.

In prometheus, setting ENABLE_ALERT_EMAILS would add an opsfile
that removes the email_configs from the alert manager.