Use cf-deployment.yml manifest to check buildpacks
We had a small script to check that all the expected buildpacks
defined in the manifest are deployed, but we do not install
the buildpacks using manifests anymore.

We will replace the manifest used for reference, and pull from
the upstream cf-deployment.yml. We expect the same buildpacks
are installed.