Fix creation of logit-syslog-drain for new envs.
This task was running in the same aggregate block as the task that
creates the space it uses - it was therefore erroring on a new
deployment because the public space didn't exist.

Combining it into the create-orgs task prevents this race.

I chose not to simply move it into the next aggregate block because that
contains app deployment tasks that require this service to be present.