Separate cloud-config generation from cf-manifest
The cloud-config is shared by all deployments using the bosh director,
so it shouldn't be coupled to the cf manifest.

Separating it out makes it easier to see what it's dependencies are. As
part of this change, I've stopped passing in an env-specific vars file
because the cloud-config doesn't use any values from it, so it was
unnecessary.