Make the GorouterLatency* alerts less noisy.
These alerts are not really useful as any long running requests can skew the
results easily. In the future we have to try to get a more useful metrics, e.g.
avg response time for only 95% of the requests and base this alert on that.

Until then we'll look at the minimum latency and we'll only alert if all
routers report a high average latency for at least 15 minutes.