Rename and enable backups for tiny redis plan
AWS now support backups for unclustered t2 class Redis instances[1],
whereas they didn't when we first introduced them (in 2b9da87). This
removes the need to have separate clustered and unclustered plans.

This change renames the tiny-unlcustered-3.2 plan to tiny-3.2 to fit
with the newly created medium plan (and to reflect that unclustered is
the default). This also enabled backups for this plan now that they're
supported.

I've also marked the tiny-clustered-3.2 plan as depracated - there's no
longer a need for this as a separate plan, and it has issues with some
clients (notably Sidekiq) that don't work with a clustered mode redis. I
elected to merely mark it as deprecated in the description at this stage
because there are lots of people using it at present (26 instances exist
on prod at the moment). We'll send out an announcement, and revisit
whether we can remove/hide it later.

[1]https://aws.amazon.com/about-aws/whats-new/2018/11/amazon_elasticache_for_redis_adds_auto-failover_and_backup_restore_support_for_t2_nodes/