Add medium-ha redis plan using m5.large and multi-az
DiT have requested larger Redis instances to be available in the London
region. They've also asked for multi-node instances to enable failover
in the event of a failure.

This adds a new plan using a cache.m5.large instance with a single
replica and automatic failover enabled. This doesn't use clustered mode
because there's no need for that unless we start to shard, and it comes
with some limitations in the ability to scale instances etc.

The new plan does not include 'unclustered' in the name as that should
be the default, and a clustered instance should be a special-case.