Configure buildpack versions with YAML
Updating the pipeline every time we need to update the buildpacks is not very sustainable, and is quite confusing to do without missing something.

`config/buildpacks.yml` is a configuration file that describes the state of our buildpacks, with each described as follows:

```yaml
- name: binary_buildpack
  repo_name: binary-buildpack
  stack: cflinuxfs2
  version: v1.0.27
  sha: 90e04df12df78f6d06ef127aa9992f505b7e51e21cb5ac97b68bc7e742193635
  filename: binary-buildpack-cflinuxfs2-v1.0.27.zip
  url: https://github.com/cloudfoundry/binary-buildpack/releases/download/v1.0.27/binary-buildpack-cflinuxfs2-v1.0.27.zip
  dependencies: []
```

Concourse will now run `parse_buildpacks_and_run_uploader.rb` which parses buildpacks.yml, and runs the existing `upload-buildpack.sh` script to download from GitHub and upload to cloudfoundry.