Remove bosh-cleanup task.
We've seen this conflicting with the rotate-*-encryption-keys jobs when
trying to get a lock in the bosh director.

We've also seen this break the prometheus deploy job (which runs in
parallel to this)  because it deletes the prometheus releases between
them being uploaded, and the deployment happening.

Given it's not essential to run the cleanup, this removes it.