Change cells to `r5.xlarge` EC2 instance type
We want to utilise Reserved Instances to reduce our EC2 costs. We want
to reserve the newest type of instances so that we are investing
effectively.

Our cells are a subset of our servers and are configured to run tenant
CF apps in containers. They represent a large portion of our EC2 costs
and so using Reserved Instances for them is a quick win.

Until now our cells have used the `r4.xlarge` EC2 instance type. This
commit changes us to use the `r5.xlarge` EC2 instance type. These new
instances have more memory (32GiB rather than 30.5GiB [1]) and so this
commit updates where that number is used.

[1] https://www.ec2instances.info/?min_memory=20&min_vcpus=4&region=eu-west-2&selected=r4.xlarge,r5.xlarge