Rename parameter group for shared_preload_libraries
There's already a manually created parameter group called
rdsbroker-postgres95-prod-pg-stat-statements in prod Ireland which was
conflicting with this. This group has instances using it, so can't be
deleted.

I initially attempted to import this into the tfstate, but terraform
needs to update the description, and AWS doesn't allow that, so
terraform still attempts to delete the group. It's therefore necessary
to use a different name for the group.