Set firehose.doppler.idle_timeout to enable re-establishing firehose connection
By default, there is no idle timeout set for the firehose exporter's connection
to the firehose web socket. During a recent incident, a network partition made
it appear to the exporter that the firehose was still connected but not sending
data. When the network was restored, the firehose exporter required restarting
manually in order for it to re-establish a connection with the firehose.

The doppler.idle_timeout setting is poorly documented, but code inspection
shows it causes timeouts to corrupt the connection state and cause
the connection to be re-established.

This commit sets the timeout to 5m to match the validity of the metrics,
so that a connection is re-established as the metrics expire in the event of
failure.