Remove ignored vcap_password from cloud-config
We were attempting to set the vcap user password through the
cloud-config, but according to the docs[1], vm_types in cloud-config
only take a name and cloud_properties block, and this extra env key was
being ignored. This could be verified by logging onto a VM, and
verifying that the password for the vcap user wasn't set (the field was
set to '*' in /etc/shadow).

This change removes the only usage of cf-secrets in the cloud-config, so
this also removes cf-secrets from the generation script etc.

[1]https://bosh.io/docs/cloud-config/#vm-types