Remove global workdir from cf manifest tests.
The coupling between the manifest helpers and the global spec_helper was
non-obvious, and made it harder to track the logic. It also was a single
dir that is shared between all the tests, leaving the opportunity for
one test to have side-effects on another.

This removes the global workdir, and instead makes it a local concern of
the render_manifest method, which simplifies things, and means that it
can be cleaned up with an ensure block.