Fix the RouterTotalRoutesDrop alert
The existing expresion was causing the following error in Prometheus:

> Error executing query: many-to-many matching not allowed: matching
> labels must be unique on one side

In fact because we're using the same series on both sides of the /
operator, there's no need for the on() group_left() functions at all.

I initially fixed this by adding 'bosh_job_name,bosh_job_id' to the on()
function, but then realised that the whole thing was unnecessary.