Remove the avg_over_time from RouterTotalRoutesDrop
We believe this was initially added in order to ensure that the alert
fired for at least 30 mins so that it would be noticed on the dashboard
etc.  We don't believe this is a good reason to do this, and it makes it
much harder to understand what this is doing.

Removing the avg_over_time also means that we no longer to combine to
functions that take a range vector and output an instant vector, which
in turn means that we don't need a record to achieve this. I've
therefore removed both records here, which will make it easier to see
what this is doing when looking at it in the Prometheus UI.