Improve DNS resolution working alert.
During the recent issues in staging where DNS resolution wasn't working
on one of the API VMs, this alert didn't fire quickly enough / at all.

This updates it alert to fire if more than 10% of the probes failes in
the last 15mins.

As part of this I've also removed the separate Warning and Critical
alert as they haven't proved helpful.