Stop specifying Postgres10 minor version for broker
The Postgres versioning scheme has changed starting with Postgres 10[1],
and it now only uses a 2 part version number instead of the 3 part
number that was used for 9.x and earlier.

For Postgres 9.x we're only specifying the first 2 parts so that we
don't have conflicts when RDS auto-upgrades the minor versions. We want
the same behaviour for Postgres 10.x, so we need this to only specify on
version part.

With this change, the plan names no longer reflect reality as RDS will
use postgres 10.6. We'l rename these in a separate PR becuase this
change will require notifying tenants.

[1]https://www.postgresql.org/support/versioning/