Use zero-downtime-push for Aiven broker.
instead of blue-green-deploy. This makes it consistent with all the
other apps we push from this pipeline. It also means that we don't need
to manually cleanup the old app.

It has another advantage in that it renames the existing app before
pushing the new app, meaning the new app is initially pushed with the
correct name. This means that it will show up in billing with this name
instead of the -new variant.