Stop using our fork of capi-release
We forked this to work around some compilation issues we were having
(see 71ec3dc and 22cf02d).

We've since tracked the issue down to a problematic interaction between
light stemcells and the bosh compiled package cache[1]. As a result
we've disabled the compiled package cache[2], which means we no longer need
to use this forked release.

[1]https://github.com/cloudfoundry/bosh/issues/2118
[2]https://github.com/alphagov/paas-bootstrap/commit/4d5be6a