Charge all apps by the second
The `ceil()` in the formula meant that every app was being charged for
one hour at a minimum, meaning that if an app was autoscaled for 30
seconds, they would still be charged for an hour.

This change ensures that the apps are now just charged for the seconds
they are active for.