Stemcell upgrade to 170.38 as part of cf7.5.0 upgrade
Whilst cf7.5.0 (and .6.0) specify a 250-series stemcell, it's failing for us at
the `bosh create-env` stage with both 250.9 and 250.17. This implies we have a
problem unrelated to the upgrade so, given that 170.38 has all the current
CVE fixes listed at [1], we're going to push forward without a major stemcell
bump.

This requires us to amend our stemcell version unit test, as cf-deployment is
ahead (250.x) of us (170.x). We've chosen to assert that **iff** cf-d and our
manifest are on the same *major* version, then the same test applies - we
mustn't be behind cf-d. However, if we've chosen to remain a major stemcell
release behind, then the test no longer applies. This was felt to be the safest
route forward which also allows for a major stemcell series upgrade to be
delayed.

[1] https://www.cloudfoundry.org/foundryblog/security-advisory/