Upgrades CF from cf-deployment 6.10.0 to 7.5.0
Notes
=====

The upgrade from v6.10.0 to v7.5.0 [1] is relatively small in terms of changes
impacting us. It's mainly a major upgrade because of the breaking change of
swapping cflinuxfs2 and 3 as the default stack.

Most package upgrades were small, and in a handful of cases we
were ahead of the upstream already (e.g. because of CVEs).

v7.1.0 introduced the nginx and R buildpacks, which we have disabled as part of
this upgrade. They will be reinstated and managed as part of our normal
buildpack upgrade stories.

This upgrade keeps us on track to remove cflinuxfs2 at the end of March, as
communicated to tenants.

Working notes available at [2].

Version bumps
=============

- bpm 1.0.0 -> 1.0.2
- cf-syslog-drain 8.1 -> 9.0
- cf-smoke-tests 40.0.44 -> 40.0.45
- diego 2.25.0 -> 2.28.0
- garden-runc 1.17.2 -> 1.18.2
- loggregator 104.4 -> 105.0
- statsd-injector 1.5.0 -> 1.7.0
- loggregator-agent 3.2 -> 3.5
- log-cache 0.1.3 -> 0.1.4 (fork maintained by us, of upstream 2.1.1 + pagination
fixes)
- ubuntu-xenial 170.30 -> 170.38

Important changes
=================

* Removed R and nginx buildpacks in
   `manifests/cf-manifest/operations.d/240-cf-disable-buildpack-installation.yml`
* Migrated from using PaaS maintained fork of CAPI package to upstream package

Issues
======

- API availability tests failed 1/10000: [3]

```
Failed to connect to Cloud Foundry API: Error getting token: oauth2: cannot
fetch token: 502 Bad Gateway
Response: 502 Bad Gateway: Registered endpoint
failed to handle the request.
```

[1] When we began the upgrade, v7.6.0 had only been released for a couple of
days and we felt it was safer to upgrade to v7.5.0 as a result.
[2] https://hackmd.cloudapps.digital/4MDiVg_PSEe6PJD3TejHdg?view
[3] https://deployer.modeerow.dev.cloudpipeline.digital/teams/main/pipelines/create-cloudfoundry/jobs/api-availability-tests/builds/85