Upgrade log-cache to 2.1.1+fixes
This is v2.1.1 + a backported fix for a 2nd pagination issue.

This first issue was fixed when upstream merged the contents of our PR
https://github.com/alphagov/paas-log-cache-release/pull/1.

A second, different fix, only made its way into log-cache 2.2.0, and is the reason
for this fork still being used by ourselves. Its contents is in
https://github.com/alphagov/paas-log-cache-release/pull/2.