Post-deploy buildpack check compares reality with buildpacks.yml
Previously, we compared the upstream `cf-deployment.yml` set of buildpacks
against reality. This isn't valid because (a) it ignores any opsfile
operations that might have altered the manifest and (b) we don't deploy
buildpacks based on the contents of that file anyway.

Our SPOT for buildpacks is `config/buildpacks.yml` - this commit asserts that
reality matches the contents of that file exactly.

We extend the buildpack-name test to include the buildpack's stack, now that
cflinuxfs3 is a thing.