Add 3 more cells in London
We are currently triggering / hovering close to triggering the alarm in London for the available memory on cells. This means we need to add more cells.

`rep_memory_capacity_pct:avg5m` is sitting around 32.8%, which is less than the threshold of 33.

As we require the number of instances to be divisible by #AZs, we scale to 18 cells.