Add Ruby 2.5 matrix build to Travis
When Travis began resolving 2.5 to 2.5.4 in [1], the build began failing for
reasons other than a test failure. This commit adds a second build to the
matrix, pinned to 2.5, which is allowed to fail. This should allow us to see
when the latest 2.5.x build is safe to use, without failing the build in the
meantime.

Matrix builds aren't carried out consecutively, and we don't believe it doubles
the build time.

[1] https://travis-ci.org/alphagov/paas-cf/builds/506785611