Estimate Aiven cost at the end of the month
The value of `paas_aiven_estimated_cost_pounds` is the current spend for
the month according to Aiven, so it goes up in a saw tooth from the
beginning of the month.

This uses a linear approximation that the cost for the month will be the
cost for the last 24 hours multiplied by 30 (for the days in the month).

The delta function just naively subtracts the value at the beginning of
the time period from the value at the end. You might think that the
`increase()` function would be a better fit, but sadly that only works
for counters that always strictly increase over time or reset to zero -
this one occasionally drops by a few pounds, which confuses prometheus.

For the first day of each month the value of
`delta(paas_aiven_estimated_cost_pounds[24h])` will be massively
negative as the counter resets to zero.

Previously we were just alerting when our spend for a month crossed a
particular threshold, which tended to happen towards the end of every
month.