Expose S3 buckets count metric
With the new S3 broker in private beta, we want to monitor the number of
buckets in the account compared to the account limit on buckets so that we can
request an increase preemptively.

This commit creates a new metric reader which calls the S3 API to list all
buckets, counts them, and exposes the number as a metric.