Enable noisy-neighbor-nozzle for CF
Recently we have had issues where the logging stack has been breaking, it has been difficult to pin this down to a specific cause. In order to gain more insight into what is happening I am electing to deploy this nozzle. in order to use it you will need the log-noise cf cli plugin installed `cf install-plugin -r CF-Community "log-noise"` and then issue `cf log-noise` to see the top log producing apps on the platform.

This is a stateless app and keeps 60 minutes of data in it's ring buffer.