Make the CloudControllerLatencyByGorouter alert less sensitive to spikes
We had a lot of false alerts, mostly because some /v2/apps queries take several
seconds to return a result.

We'll only look at a 5 minute average and this value has to be constantly over
500ms for 15 minutes to trigger an alert.

I removed the test as this is a trivial expression, and imho it's a waste of
time to maintain it.