Increase maximum app disk usage to 4GB
One of our tenants wants to push a Docker image that is 2.2GB in size.
Cloud Foundry limits the disk usage of apps to 2GB by default, and so
they have not been able to release their app on the PaaS.

This commit speculatively increases that limit to 4GB, so that we have
a PR to discuss the change on.