Add in a tool to pull email addresses of tenants
We have added in a tool to gather the email addresses of all tenants in order to send highly important emails. This is due to the lack of use of the PaaS-Annoounce email group. There are two levels of data that this tool provides. The first is that when run without any flags it will obtain all the email addresses of space developers. When run with the `CRITICAL=true` flag it will pull all email addresses for tenants. These are passed through deduplication in advance of output in order to not send multiple emails. Email aliasing has not been addressed as part of this work, we also considered that sending multiple emails may well still have validity due to people using filtering on the output. e.g. first.last+paas@domain.com may be used for their inbox management.