Revert "Enable s3 broker separately to registering it"
This reverts commit 313309ed73a79ec1bf8a1e89504c64b3f50122d1. That
commit moved enabling the `aws-s3-bucket` service to a later place in
the pipeline, because the pipeline was only enabling for a particular
space and there was a race condition with the creation of that space.

We are about to enable the `aws-s3-bucket` service for everyone, and
so that is no longer necessary.