Remove cflinuxfs2 buildpacks from config
Although this file is auto-generated by `scripts/update_buildpacks.sh`,
it uses its current state as input. This means removing the buildpacks
is enough to remove them from the generated output. You can confirm this
by running the script and observing that the buildpacks are not
re-added.

This commit is different from 163c6f9e since the file still contains a
bunch of references to cflinuxfs2 in the specification of which stacks
dependencies work on. This is just an unfortunate artifact of the way
the script to generate buildpacks.yml works. We shouldn't remove them or
they'll just get added back the next time we generate this file.