Clarify alert when multiple Stemcells are in-use
We recently added an alert to warn us when more than one BOSH Stemcell version is in use. This alert is helpful because it is easy to forget that Stemcell upgrades must be performed in both https://github.com/alphagov/paas-bootstrap and https://github.com/alphagov/paas-cf.

The alert went off for the first time to warn of this problem, but the `paas-bootstrap and paas-cf are not in sync` description may not have been specific enough to explain the problem. This commit clarifies the message further to make the alert more effective.