Add output to a secret rotation script
The `rotate-vars-store-secrets.rb` script changed by this commit is a
little obscure. It is used in several different ways by the secret
rotation tasks in our pipelines.

One of the weirdest things about this script is that it does not
actually rotate secrets. It moves, deletes or blanks existing secrets,
but it does not generate new values. I'm not sure what *does*, but I
suspect the answer is BOSH.

Regardless, it will help our use of this script if it describes what
it does. The output is not sensitive: none of the values are output,
just the names of the variables, and those names are in our manifests
anyway.