Refactor metric tags to a slice of structs
Previously, metric tags were a slice of strings, which were required to be
"key:value" pairs. This wasn't clear from the interface alone, and was only
obvious on seeing the error message in the prometheus reporter (which made the
assumption about string format). In this commit, I refactored the tags to be a
slice of the MetricTag struct, and added some stringer methods. This should
make the interface clearer, and this particular error avoidable in future.