Show DNS Resolution times on User Impact dashboard
In an incident on 2019-05-20, we took awhile to realise that network
problems on one of the VMs running Cloud Controller was the cause of
elevated API error rates. We have seen this problem before. As this is
a repeated and open issue it is definitely worth trying to improve our
response next time.

This commit adds DNS Resolution times to the User Impact dashboard.
It is highly visible in the team area and commonly looked at
out-of-hours to know if there is a problem. Despite the name, the User
Impact dashboard is not directly about user-facing problems.

The JSON is copied from manually editing this dashboard in `prod-lon`.