Fetch average of last 10 minutes, not since 2012
Previously, the CloudFront metrics collection specified January 2012 as
a start date, and the current datetime as the end date. Although this
does show us the change over time, the sensitivity to change will vary
wildly, depending on the age of the cdn-route service. Additionally,
every collection period will refer to a different length of time as time
goes on, making comparison difficult / fruitless.

Instead, just collect an average of the last 10 minutes' data from
CloudWatch - this way we can compare apples to apples, not to apple
trees.