Log cache to use TLS behind ELB
In log-cache-release 2.2.0 TLS is mandatory after whatever load balancer
is configured.

We set this in our terraform, as that is what manages the ELB.

Update the ELB to use TLS for backend connections and for healthchecks.
By default ELBs trust whatever cert is given to them when negotiating a
connection.

Co-authored-by: Lee <lee.porte@digital.cabinet-office.gov.uk>
Co-authored-by: Toby <toby.lornewelch-richards@digital.cabinet-office.gov.uk>