Adjust user impact dashboard to be more legible
The graphs showing 2xx, 4xx and 5xx were scaled according to the highest
value (usually the 2xx requests), and so 4xx and 5xx were not easily
visible.

The request type with the most requests has now been moved to the right
axis, which means that 4xx and 5xx are autoscaled separately.

Also, the 2xx results were a little too BIG and BOLD, which drew
attention away from the 'point' of the graph (showing the problems).
These are now much less bold, so the eye is drawn to the issues.

FINALLY, Colours for the request graphs have been set manually, so
they merge together less.