Add initial SLI dashboard
About this dashboard
--------------------

We've decided to measure "uptime" based on metrics from the CloudFront CDN. CloudFront sits on the outside the platform, so any errors due to the platform will be seen in the cloud front metrics.

At the time of writing there are about 40 CloudFront CDNs on the platform (across London and Ireland). They cover both production and non-production services.

Our SLI should not be affected if a few services have high error rates - this probably indicates an issue with the individual applications.
If a large number of services have high error rates at the same time we should assume there's a problem with the platform.

Initially our application uptime SLI is:

```
The average (mean) percentage of 5XX responses from CloudFront, ignoring the 5 distributions with the highest percentage of errors
```

Known issues
------------

* If more than 5 services have high error rates due to non-platform issues (e.g. bad application code), this will unfairly affect our SLI
* If a platform issue causes high error rates for fewer than 5 services our SLI will not be affected
* The `5xx_errorrate` metric is a percentage, so low-traffic services with a few requests per minute are weighted equally with high-traffic services which have many requests per second

References
----------

See [the pivotal tracker story to implement this dashboard](https://www.pivotaltracker.com/story/show/165379426)