Make metrics/tlscheck tests less flakey for badssl
badssl.com hosts a number of bad certificates which can be used by test
suites. we depend on it for our metrics tests.

recently they have been returning an expired, AND self-signed
certificate for self-signed.badssl.com.

our tests expect this not to be expired, but to be self signed. our
tests then fail because the code (correctly) returns no-error and zero
days until expiry

if we skip verification, we ignore if the certificate is expired, which
is the desired behaviour for this test

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>