Script for adding existing UAA users to paas-accounts
This script is a one-off, but we're raising it on a separate branch so we can
review it in our normal workflow.

The script fetches all UAA users, and either creates or updates them in
paas-accounts as appropriate. If the username is a valid email, we populate the
email field. Otherwise, it is null.