Refactor API availability tests to obey % maximum
We often have to re-kick-off the API availability tests because of a
single Failure, or small number of warnings. Especially when the deploys
take a long time (e.g. Diego cell rolling).

A single failure is not tolerated (e.g. a 502 from UAA), we often see
15000 total executions and a single 502 and this causes a test failure,
even though it is 99.99% reliable.

99.95% is a good enough reliability threshold for the API during
deployments.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>