Allow subshell scope logins for dev envs
In the same spirit as `scripts/cf_subshell_scoped_login.sh`, this commit adds
an option to `scripts/cf_admin_login_dev` to allow users to scope the login to
a subshell. It uses the same method to do that as the former.

This is intended as a quality of life improvement, for those times when devs
are working across multiple spaces and the global nature of the `cf` tool is a
hindrance.