Add "group" to pricing plans along with name
Sometimes we want to ask a question like:

"How much has this organisation spent on postgres?"
"How much has this organisation spent on compute (including staging and tasks)?"

At the moment this is more difficult than it needs to be because we have
to add up all the various postgres plans the organisation might be
using. This currently involves doing string processing to work out which
plans are postgres based on which names start with "postgres", which
feels super hacky.

The plan would be to add a group field to the pricing plans in the DB,
and then expose this through the /pricing_plans API in paas-billing.
This would allow us to remove the code that does string processing on
"name".

This commit adds the following groups:

* aws-s3-bucket
* cloudfront
* compute (includes app, staging and task)
* elasticsearch
* mongodb
* mysql
* postgres
* prometheus
* redis
* unknown

I think it's best to add the groups to the existing plans so they're
there for all history. The [paas-billing code](https://github.com/alphagov/paas-billing/blob/94f16d71fda09ae19278e213e5e93743231deadb/eventstore/store.go#L210-L224)
replaces all the pricing plans when it starts up. Once that code is
updated to include groups (separate PR to be created later) we'll be
able to start exposing them in queries.

Review wise, firstly think about whether this is a good idea in general
(is there some complexity I'm missing?). If so, is "group" the best name
for the new key, or would something else ("category"?) be better? Do the
groupings I've come up with make sense?