Reduce Ireland from 24 to 18 cells
We maintain enough cells to satisfy our tenant demand for app memory
and our reliability goals [1]. Since the evening of July 10th, demand
in Ireland has decreased to 16 cells at night and 17 cells in the day.

This commit reduces us to the nearest multiple of 3 above that demand,
which is 18. We expect demand in Ireland to continue to fall as
tenants migrate to our London environment hence this aggressive
reduction.

[1] https://team-manual.cloud.service.gov.uk/architecture_decision_records/ADR021-cell-capacity-assignment-2/